Added swim history table to swimmer page
Click on an event in the PB list, and a history of swims for that event
will be created.
Got rid of redundant 'Race' class in personalbests.js and unified with
the 'Swim' class from swim.js.
Teased out the code to create a race time <td> element into a free
function.
Fixed a massive bug caused by JavaScript's ridiculous Date convention of
numbering days from 1, but months from 0.