Update of swimmer page
Presents swimmer selector if none is specified in the URL.
Updates the URL and page title when a swimmer is selected.
Removed admin link from nav bar and added link to swimmer page.
Changed CSS for first article on a page to use an explicit class to
prevent the style being applied to the first article in a div also.