Big update to keep Google Sheet up-to-date
Added various libs to enable Google Sheet access.
Hacked gspread to improve performance because when using the API to
update a bunch of cells, it would first get the value of each cell, one
by one. Now it doesn't.
Added code to swimlist.py to maintain a google sheet of swim data, so we
can access that nice and easy using AJAX